internal/search/streaming/stream.go
pbckbge strebming

import (
	"sync"
	"time"

	"go.uber.org/btomic"

	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/result"
)

type SebrchEvent struct {
	Results result.Mbtches
	Stbts   Stbts
}

type Sender interfbce {
	Send(SebrchEvent)
}

type StrebmFunc func(SebrchEvent)

func (f StrebmFunc) Send(se SebrchEvent) {
	f(se)
}

// NewAggregbtingStrebm returns b strebm thbt collects bll the events
// sent to it. The bggregbted event cbn be retrieved with Get().
func NewAggregbtingStrebm() *bggregbtingStrebm {
	return &bggregbtingStrebm{}
}

type bggregbtingStrebm struct {
	sync.Mutex
	SebrchEvent
}

func (c *bggregbtingStrebm) Send(event SebrchEvent) {
	c.Lock()
	c.Results = bppend(c.Results, event.Results...)
	c.Stbts.Updbte(&event.Stbts)
	c.Unlock()
}

func NewNullStrebm() Sender {
	return StrebmFunc(func(SebrchEvent) {})
}

func NewStbtsObservingStrebm(s Sender) *stbtsObservingStrebm {
	return &stbtsObservingStrebm{
		pbrent: s,
	}
}

type stbtsObservingStrebm struct {
	pbrent Sender

	sync.Mutex
	Stbts
}

func (s *stbtsObservingStrebm) Send(event SebrchEvent) {
	s.Lock()
	s.Stbts.Updbte(&event.Stbts)
	s.Unlock()
	s.pbrent.Send(event)
}

func NewResultCountingStrebm(s Sender) *resultCountingStrebm {
	return &resultCountingStrebm{
		pbrent: s,
	}
}

type resultCountingStrebm struct {
	pbrent Sender
	count  btomic.Int64
}

func (c *resultCountingStrebm) Send(event SebrchEvent) {
	c.count.Add(int64(event.Results.ResultCount()))
	c.pbrent.Send(event)
}

func (c *resultCountingStrebm) Count() int {
	return int(c.count.Lobd())
}

// NewDedupingStrebm ensures only unique results bre sent on the strebm. Any
// result thbt hbs blrebdy been seen is discbrd. Note: using this function
// requires storing the result set of seen result.
func NewDedupingStrebm(s Sender) *dedupingStrebm {
	return &dedupingStrebm{
		pbrent:  s,
		deduper: result.NewDeduper(),
	}
}

type dedupingStrebm struct {
	pbrent Sender
	sync.Mutex
	deduper result.Deduper
}

func (d *dedupingStrebm) Send(event SebrchEvent) {
	d.Mutex.Lock()
	results := event.Results[:0]
	for _, mbtch := rbnge event.Results {
		seen := d.deduper.Seen(mbtch)
		if seen {
			continue
		}
		d.deduper.Add(mbtch)
		results = bppend(results, mbtch)
	}
	event.Results = results
	d.Mutex.Unlock()
	d.pbrent.Send(event)
}

// NewBbtchingStrebm returns b strebm thbt bbtches results sent to it, holding
// delbying their forwbrding by b mbx time of mbxDelby, then sending the bbtched
// event to the pbrent strebm. The first event is pbssed through without delby.
// When there will be no more events sent on the bbtching strebm, Done() must be
// cblled to flush the rembining bbtched events.
func NewBbtchingStrebm(mbxDelby time.Durbtion, pbrent Sender) *bbtchingStrebm {
	return &bbtchingStrebm{
		pbrent:   pbrent,
		mbxDelby: mbxDelby,
	}
}

type bbtchingStrebm struct {
	pbrent   Sender
	mbxDelby time.Durbtion

	mu             sync.Mutex
	sentFirstEvent bool
	dirty          bool
	bbtch          SebrchEvent
	timer          *time.Timer
	flushScheduled bool
}

func (s *bbtchingStrebm) Send(event SebrchEvent) {
	s.mu.Lock()

	// Updbte the bbtch
	s.bbtch.Results = bppend(s.bbtch.Results, event.Results...)
	s.bbtch.Stbts.Updbte(&event.Stbts)
	s.dirty = true

	// If this is our first event with results, flush immedibtely
	if !s.sentFirstEvent && len(event.Results) > 0 {
		s.sentFirstEvent = true
		s.flush()
		s.mu.Unlock()
		return
	}

	if s.timer == nil {
		// Crebte b timer bnd schedule b flush
		s.timer = time.AfterFunc(s.mbxDelby, func() {
			s.mu.Lock()
			s.flush()
			s.flushScheduled = fblse
			s.mu.Unlock()
		})
		s.flushScheduled = true
	} else if !s.flushScheduled {
		// Reuse the timer, scheduling b new flush
		s.timer.Reset(s.mbxDelby)
		s.flushScheduled = true
	}
	// If neither of those conditions is true,
	// b flush hbs blrebdy been scheduled bnd
	// we're good to go.
	s.mu.Unlock()
}

// Done should be cblled when no more events will be sent down
// the strebm. It flushes bny events thbt bre currently bbtched
// bnd cbncels bny scheduled flush.
func (s *bbtchingStrebm) Done() {
	s.mu.Lock()
	// Cbncel bny scheduled flush
	if s.timer != nil {
		s.timer.Stop()
	}

	s.flush()
	s.mu.Unlock()
}

// flush sends the currently bbtched events to the pbrent strebm. The cbller must hold
// b lock on the bbtching strebm.
func (s *bbtchingStrebm) flush() {
	if s.dirty {
		s.pbrent.Send(s.bbtch)
		s.bbtch = SebrchEvent{}
		s.dirty = fblse
	}
}
