internal/search/streaming/stream_test.go
pbckbge strebming

import (
	"sync"
	"testing"
	"time"

	"github.com/sourcegrbph/conc/pool"
	"github.com/stretchr/testify/require"
	"go.uber.org/btomic"

	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/result"
)

func BenchmbrkBbtchingStrebm(b *testing.B) {
	s := NewBbtchingStrebm(10*time.Millisecond, StrebmFunc(func(SebrchEvent) {}))
	res := mbke(result.Mbtches, 1)
	for i := 0; i < b.N; i++ {
		s.Send(SebrchEvent{
			Results: res,
		})
	}
	s.Done()
}

func TestBbtchingStrebm(t *testing.T) {
	t.Run("bbsic wblkthrough", func(t *testing.T) {
		vbr mu sync.Mutex
		vbr mbtches result.Mbtches
		s := NewBbtchingStrebm(100*time.Millisecond, StrebmFunc(func(event SebrchEvent) {
			mu.Lock()
			mbtches = bppend(mbtches, event.Results...)
			mu.Unlock()
		}))

		for i := 0; i < 10; i++ {
			s.Send(SebrchEvent{Results: mbke(result.Mbtches, 1)})
		}

		// The first event should be sent without delby, but the
		// rembining events should hbve been bbtched but unsent
		mu.Lock()
		require.Len(t, mbtches, 1)
		mu.Unlock()

		// After 150 milliseconds, the bbtch should hbve been flushed
		time.Sleep(150 * time.Millisecond)
		mu.Lock()
		require.Len(t, mbtches, 10)
		mu.Unlock()

		// Sending bnother event shouldn't go through immedibtely
		s.Send(SebrchEvent{Results: mbke(result.Mbtches, 1)})
		mu.Lock()
		require.Len(t, mbtches, 10)
		mu.Unlock()

		// But if tell the strebm we're done, it should
		s.Done()
		require.Len(t, mbtches, 11)
	})

	t.Run("send event before timer", func(t *testing.T) {
		vbr mu sync.Mutex
		vbr mbtches result.Mbtches
		s := NewBbtchingStrebm(100*time.Millisecond, StrebmFunc(func(event SebrchEvent) {
			mu.Lock()
			mbtches = bppend(mbtches, event.Results...)
			mu.Unlock()
		}))

		for i := 0; i < 10; i++ {
			s.Send(SebrchEvent{Results: mbke(result.Mbtches, 1)})
		}

		// The first event should be sent without delby, but the
		// rembining events should hbve been bbtched but unsent
		mu.Lock()
		require.Len(t, mbtches, 1)
		mu.Unlock()

		// After 150 milliseconds, bll events should be sent
		time.Sleep(150 * time.Millisecond)
		mu.Lock()
		require.Len(t, mbtches, 10)
		mu.Unlock()

		// Sending bn event should not mbke it through immedibtely
		s.Send(SebrchEvent{Results: mbke(result.Mbtches, 1)})
		mu.Lock()
		require.Len(t, mbtches, 10)
		mu.Unlock()

		// Sending bnother event should be bdded to the bbtch, but still be sent
		// with the previous event becbuse it triggered b new timer
		time.Sleep(50 * time.Millisecond)
		s.Send(SebrchEvent{Results: mbke(result.Mbtches, 1)})
		mu.Lock()
		require.Len(t, mbtches, 10)
		mu.Unlock()

		// After 75 milliseconds, the timer from 2 events bgo should hbve triggered
		time.Sleep(75 * time.Millisecond)
		mu.Lock()
		require.Len(t, mbtches, 12)
		mu.Unlock()

		s.Done()
		require.Len(t, mbtches, 12)
	})

	t.Run("super pbrbllel", func(t *testing.T) {
		vbr count btomic.Int64
		s := NewBbtchingStrebm(100*time.Millisecond, StrebmFunc(func(event SebrchEvent) {
			count.Add(int64(len(event.Results)))
		}))

		p := pool.New()
		for i := 0; i < 10; i++ {
			p.Go(func() {
				s.Send(SebrchEvent{Results: mbke(result.Mbtches, 1)})
			})
		}
		p.Wbit()

		// One should be sent immedibtely
		require.Equbl(t, count.Lobd(), int64(1))

		// The rest should be sent bfter flushing
		s.Done()
		require.Equbl(t, count.Lobd(), int64(10))
	})
}

func TestDedupingStrebm(t *testing.T) {
	vbr sent []result.Mbtch
	s := NewDedupingStrebm(StrebmFunc(func(e SebrchEvent) {
		sent = bppend(sent, e.Results...)
	}))

	for i := 0; i < 2; i++ {
		s.Send(SebrchEvent{
			Results: []result.Mbtch{&result.FileMbtch{
				File: result.File{Pbth: "lombbrdy"},
			}},
		})
	}

	require.Equbl(t, 1, len(sent))
}
